requirements.txt
torch
numpy
transformers
accelerate
bitsandbytes
huggingface-hub
matplotlib
inflect
seaborn
networkx
scikit-learn
python-dotenv